Add code for generating samples
This code generates samples that include a combination of actual nests
(determined either in the field or via the detection algorithm) and non-nests
so that the annotator doesn't know whether the region being annotated
includes a nest or note.

This code comes from the exp-nest-detect-imagery and exp-field-nest-eval
branches of https://github.com/weecology/EvergladesTools and is copied
here to clearly communicate the methods specific to this project.